Fixed circular imports and outdated script
